[refactor] 필요한 어노테이션 추가 및 반환 타입 명시
